Fix comparisons between signed and unsigned values
In most cases, this was caused by using a signed integer type for a
loop/array index; we probably ought to use size_t in these cases just
so there's no unexpected behavior from comparing signed/unsigned
types. In some other cases, I explicitly cast the signed value to
unsigned before comparing; we might need to consider whether the
signed value actually needs to be signed in these cases.

+++ src/Autonomous.cpp
@@ -312,7 +312,7 @@ void Autonomous::updateLandmarkInformation(const transform_t &invTransform)
 	// get landmark data and filter out invalid data points
 	points_t landmarks = readLandmarks();
 	printLandmarks(landmarks);
-	if (urc_targets.front().left_post_id < 0 || urc_targets.front().left_post_id > landmarks.size())
+	if (urc_targets.front().left_post_id < 0 || static_cast<unsigned>(urc_targets.front().left_post_id) > landmarks.size())
 	{
 		log(LOG_ERROR, "Invalid left_post_id %d\n", urc_targets.front().left_post_id);
 		return;
@@ -564,7 +564,7 @@ void Autonomous::autonomyIter()
   if (USE_MAP) {
       if (mapLoopCounter++ > mapBlindPeriod) {
           map.addPoints(toTransform(pose), lidar_scan, mapDoesOverlap ? 0.4 : 0);
-          mapDoesOverlap = lidar_scan.size() > mapOverlapSampleThreshold;
+          mapDoesOverlap = lidar_scan.size() > static_cast<unsigned>(mapOverlapSampleThreshold);
       }
   }
 
@@ -714,7 +714,7 @@ pose_t Autonomous::getGPSTargetPose() const
 static void printLandmarks(points_t& landmarks, int log_level){
 	std::ostringstream stream;
 	stream << "Landmarks: [";
-	for(int i = 0; i < landmarks.size(); i++){
+	for(size_t i = 0; i < landmarks.size(); i++){
 		stream << i << ": {"
 			   << landmarks[i][0] << ", "
 			   << landmarks[i][1] << ", "
